[del/#217] SettingsNavigation.kt 파일 제거
