+++ feature/mypage/src/main/java/com/napzak/market/mypage/navigation/MyPageNavigation.kt
@@ -2,18 +2,27 @@ package com.napzak.market.mypage.navigation
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.napzak.market.common.navigation.MainTabRoute
+import com.napzak.market.common.navigation.Route
 import com.napzak.market.mypage.mypage.MyPageRoute
+import com.napzak.market.mypage.setting.SettingsRoute
+import com.napzak.market.util.common.openUrl
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateToMyPage(navOptions: NavOptions? = null) =
     navigate(MyPage, navOptions)
 
+fun NavController.navigateToSettings(navOptions: NavOptions? = null) {
+    this.navigate(Settings, navOptions)
+}
+
 fun NavGraphBuilder.mypageGraph(
+    restartApplication: () -> Unit,
     navigateToUp: () -> Unit,
     navigateToMyMarket: (Long) -> Unit,
     navigateToSales: () -> Unit,
@@ -21,6 +30,7 @@ fun NavGraphBuilder.mypageGraph(
     navigateToRecent: () -> Unit,
     navigateToFavorite: () -> Unit,
     navigateToSettings: () -> Unit,
+    navigateToSignOut: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
     composable<MyPage> {
@@ -38,7 +48,23 @@ fun NavGraphBuilder.mypageGraph(
             modifier = modifier,
         )
     }
+
+    composable<Settings> {
+        val context = LocalContext.current
+
+        SettingsRoute(
+            onBackClick = navigateToUp,
+            onLogoutConfirm = restartApplication,
+            onWithdrawClick = navigateToSignOut,
+            openWebLink = { url ->
+                context.openUrl(url)
+            }
+        )
+    }
 }
 
 @Serializable
 data object MyPage : MainTabRoute
+
+@Serializable
+data object Settings : Route
